+++ src/InputFileAction.php
@@ -19,6 +19,10 @@ class InputFileAction extends ClientBaseAction
      * @var string Separator for multiple paths in textarea
      */
     public $textareaSeparator = '\n';
+    /**
+     * @var bool Put file name without path into field
+     */
+    public $nameOnly = false;
 
     /**
      * {@inheritdoc}
@@ -29,6 +33,7 @@ class InputFileAction extends ClientBaseAction
         $id = Html::encode($id);
 
         $multiple = Yii::$app->request->getQueryParam('multiple');
+        $fileParam = $this->nameOnly ? 'name' : 'url';
 
         if (!empty($multiple)) {
             $this->settings['commandsOptions']['getfile']['multiple'] = true;
@@ -36,7 +41,7 @@ class InputFileAction extends ClientBaseAction
 function (files) {
     var urls = [], separator = "{$this->separator}", el = window.opener.jQuery("#$id"), value = el.val();
     for (var i in files) {
-        urls.push(files[i].url);
+        urls.push(files[i].$fileParam);
     }
     if (el.prop("tagName").toLowerCase() == "textarea") separator = "{$this->textareaSeparator}";
     if (value) {
@@ -50,7 +55,7 @@ JSEXP;
         } else {
             $callback = <<<JSEXP
 function (file) {
-    window.opener.jQuery("#$id").val(file.url).trigger("change");
+    window.opener.jQuery("#$id").val(file.$fileParam).trigger("change");
     window.close();
 }
 JSEXP;